src/requirements.txt
Adafruit-Blinka==8.22.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-framebuf==1.6.3
adafruit-circuitpython-requests==2.0.1
adafruit-circuitpython-ssd1306==2.12.14
adafruit-circuitpython-typing==1.9.4
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.52.0
Adafruit-PureIO==1.1.11
aiohttp>=3.10.0,<3.11.0
aiosignal==1.3.1
annotated-types>=0.6.0
anyio>=4.0.0
async-timeout==4.0.3
attrs==23.1.0
caldav==1.3.9
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
fastapi>=0.103.0
frozenlist==1.4.0
h11==0.14.0
httpcore>=0.18.0
httplib2==0.22.0
httpx>=0.25.2
idna==3.4
multidict==6.0.4
ninja==1.11.1
oauthlib==3.2.2
phue==1.1
Pillow==10.0.0
protobuf==3.20.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pydantic>=2.7.4,<3.0.0
pyftdi==0.55.0
pyparsing==3.1.1
pyserial==3.5
python-dotenv==1.0.0
python-multipart==0.0.9
pyttsx3==2.90
pyusb==1.2.1
requests>=2.32.5
requests-oauthlib==1.3.1
RPi.GPIO==0.7.1
rsa==4.9
sniffio==1.3.0
dbus-python==1.3.2
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-checkpoint-postgres==3.0.2
langmem==0.0.30
langchain==1.2.0
langchain-core==1.2.6
langchain-community==0.4.1
langchain-openai==1.1.5
langchain-anthropic==1.3.0
langchain-litellm==0.3.2
litellm>=1.65.1
langsmith==0.3.45
mutagen==1.47.0
numpy==1.26.4
psycopg[binary]==3.2.4
psycopg-pool==3.2.4
SpeechRecognition>=3.11.0
spidev==3.6
spotipy==2.23.0
starlette>=0.37.0
sse-starlette>=2.0.0
tqdm==4.66.1
typing_extensions>=4.7
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
websockets>=12.0
yarl==1.9.2
gTTS==2.5.4
pygame-ce==2.5.3
pydub==0.25.1
pyudev==0.24.1
psutil==5.9.8
webrtcvad==2.0.10
